requirements.txt
numpy
librosa
moviepy
opencv-python
functools
pillow
pandas
scipy
pytube